Updated to account for @sdpalley PR
In https://github.com/vzvenyach/legalmd/pull/30, @sdpalley included excellent recommendations to improve the css.
In addition, he recommended changing the starting text to give someone more context about why this tool would be useful.
This commit incorporates those recommendations, and does it in the JSX file, so that the changes aren't lost when new features are added.